Use apollo state for query results
Uses the apollo state management capabilities, removes jotai state management and useEffects to update state.
Apollo already manages state of queries, so we can use that instead of managing state ourselves.
This now works because I've added cache mechanism to recognize `pk` as a unique identifier for each object. When an update is called, it will update the cache with the new object, which will then update the UI.