Load URI's from environment file
Instead of environment variables that are hardcoded into the build output, this will load URI's based on the hostname of the current page. This will allow us to have different URI's for different environments without having to rebuild the application.

A next step would be to publish the dist files somewhere that navigate-server can access them (like on Sonatype). So a build step for this repo is not needed for deploying navigate.